condition if pour afficher panier si rempli et afficher panier si vide

+++ javascript/order.js
@@ -5,14 +5,29 @@ const total = JSON.parse(localStorage.getItem("total"));
 // Affichage des données sur la page de confirmation
 const div = document.querySelector('div');
 
-const pTotal = document.createElement('p');
-pTotal.setAttribute("class", "p-total-order");
-pTotal.innerText = `Nous vous remercions pour votre commande d'un total de ${total} €`;
-div.appendChild(pTotal);
+if(total === null && orderId === null) {
 
-const pOrderId = document.createElement('p');
-pOrderId.setAttribute("class", "p-order-id");
-pOrderId.innerText = `Voici votre numéro de commande: ${orderId}`;
-div.appendChild(pOrderId);
+    const pNull = document.createElement('p');
+    pNull.setAttribute("class", "p-null");
+    pNull.innerText = `Votre panier est vide 😅`;
+    div.prepend(pNull);
+    
+} else {
 
-localStorage.clear();
+    const h2 = document.createElement("h2");
+    h2.innerText = "Félicitations ! 🥳"
+    div.appendChild(h2);
+
+    const pTotal = document.createElement('p');
+    pTotal.setAttribute("class", "p-total-order");
+    pTotal.innerText = `Nous vous remercions pour votre commande d'un total de ${total} €`;
+    div.appendChild(pTotal);
+
+    const pOrderId = document.createElement('p');
+    pOrderId.setAttribute("class", "p-order-id");
+    pOrderId.innerText = `Voici votre numéro de commande: ${orderId}`;
+    div.appendChild(pOrderId);
+
+    localStorage.clear(); 
+}
+
